small modify on backtester

+++ Trading.Alpha/Trading.Backtest/Data/DataCache.cs
@@ -10,9 +10,9 @@ namespace Trading.Backtest.Data
         public static DataCache Instance { get { return c.Value; } }
 
         private readonly Dictionary<DateTime, Dictionary<int, double>> volumeCache = new Dictionary<DateTime, Dictionary<int, double>>();
-        private readonly Dictionary<int, Security> securityCache = new Dictionary<int, Security>();
+        private readonly Dictionary<long, Security> securityCache = new Dictionary<long, Security>();
 
         public static Dictionary<DateTime, Dictionary<int, double>> VolumeCache { get { return Instance.volumeCache; } }
-        public static Dictionary<int, Security> SecurityCache { get { return Instance.securityCache; } }
+        public static Dictionary<long, Security> SecurityCache { get { return Instance.securityCache; } }
     }
 }
